siege-weapon/requirements.txt
future==0.17.1
iso8601==0.1.12
pyserial==3.4
PyYAML==3.13
